Add environment module to abstract app init
