+++ src/pages/home_page.rs
@@ -1,3 +1,4 @@
+#![allow(non_snake_case)]
 use dioxus::prelude::*;
 use dioxus_free_icons::icons::bs_icons::BsHouseDoorFill;
 use dioxus_router::Link;
